replace if-else with polymorphism
