+++ src/main/java/com/company/secondtask/Main.java
@@ -4,6 +4,8 @@ import com.company.secondtask.command.Command;
 import com.company.secondtask.command.MoveForwardTractorCommand;
 import com.company.secondtask.command.TurnClockwiseTractorCommand;
 import com.company.secondtask.figure.Tractor;
+import com.company.secondtask.figure.orientation.North;
+import com.company.secondtask.figure.orientation.Orientation;
 
 /**
  * Необходимо выполнить рефакторинг существующего кода - исправить все места, которые вам покажутся некрасивыми или не
@@ -17,7 +19,7 @@ public class Main {
     public static void main(String[] args) {
         Field field = new Field(5, 5);
         Position startTractorPosition = new Position(0, 0);
-        Orientation startTractorOrientation = Orientation.NORTH;
+        Orientation startTractorOrientation = new North();
         Tractor tractor = new Tractor(field, startTractorPosition, startTractorOrientation);
         Command moveForwardTractor = new MoveForwardTractorCommand(tractor);
         Command turnClockwiseTractorCommand = new TurnClockwiseTractorCommand(tractor);
