+++ src/main/java/com/company/secondtask/command/MoveForwardTractorCommand.java
@@ -2,6 +2,7 @@ package com.company.secondtask.command;
 
 import com.company.secondtask.*;
 import com.company.secondtask.figure.Tractor;
+import com.company.secondtask.figure.orientation.Orientation;
 
 public class MoveForwardTractorCommand implements Command {
 
@@ -15,15 +16,7 @@ public class MoveForwardTractorCommand implements Command {
     public void execute() {
         Orientation orientation = tractor.getOrientation();
         Position position = tractor.getPosition();
-        if (orientation == Orientation.NORTH) {
-            position.setY(position.getY() + 1);
-        } else if (orientation == Orientation.EAST) {
-            position.setX(position.getX() + 1);
-        } else if (orientation == Orientation.SOUTH) {
-            position.setY(position.getY() - 1);
-        } else if (orientation == Orientation.WEST) {
-            position.setX(position.getX() - 1);
-        }
+        orientation.moveOneCellForward(position);
         if (tractor.getField().outOfField(position)) {
             throw new TractorInDitchException("Tractor went out of field.");
         }
