+++ src/main/java/com/company/secondtask/figure/orientation/Orientation.java
@@ -1,0 +1,10 @@
+package com.company.secondtask.figure.orientation;
+
+import com.company.secondtask.Position;
+
+public interface Orientation {
+
+    Orientation turnRight();
+
+    void moveOneCellForward(Position position);
+}
